# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Local AI Relay API",
    "version": "0.1.0",
    "description": "Authenticated, OpenAI-compatible loopback API. Browser-provider readiness is runtime-gated.",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Default loopback endpoint"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/models": {
      "get": {
        "operationId": "listModels",
        "summary": "List runtime-ready models",
        "parameters": [
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all"
              ]
            },
            "description": "Use all to include registered but unready models and capability metadata."
          }
        ],
        "responses": {
          "200": {
            "description": "Model inventory.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelList"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/providers/status": {
      "get": {
        "operationId": "listProviderStatus",
        "summary": "List provider readiness evidence",
        "responses": {
          "200": {
            "description": "Provider capability records.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "object",
                    "data"
                  ],
                  "properties": {
                    "object": {
                      "const": "list"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProviderStatus"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing or invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    },
    "schemas": {
...
      "RelayMetadata": {
        "type": "object",
        "required": [
          "transport",
          "execution_style",
          "supports_sessions",
          "supports_streaming",
          "max_parallel_requests"
        ],
        "properties": {
          "transport": {
            "type": "string",
            "enum": [
              "mock",
              "browser",
              "api",
              "local"
            ]
          },
          "execution_style": {
            "type": "string",
            "enum": [
              "direct",
              "batch",
              "delegate"
            ]
          },
          "supports_sessions": {
            "type": "boolean"
          },
          "supports_streaming": {
            "type": "boolean"
          },
          "max_parallel_requests": {
            "type": "integer",
            "minimum": 1
          },
          "capability_status": {
            "type": "string"
          },
          "capability_detail": {
            "type": "string"
          },
          "capability_evidence": {
            "type": "string"
          },
          "capability_updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Model": {
        "type": "object",
        "required": [
          "id",
          "object",
          "created",
          "owned_by"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "const": "model"
          },
          "created": {
            "type": "integer"
          },
          "owned_by": {
            "type": "string"
          },
          "x_relay": {
            "$ref": "#/components/schemas/RelayMetadata"
          }
        }
      },
      "ModelList": {
        "type": "object",
        "required": [
          "object",
          "data"
        ],
        "properties": {
          "object": {
            "const": "list"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          }
        }
      },
      "ProviderStatus": {
        "type": "object",
        "required": [
          "provider_id",
          "status",
          "ready",
          "evidence_expired",
          "detail",
          "updated_at"
        ],
        "properties": {
          "provider_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "installed",
              "authenticated",
              "reachable",
              "ready",
              "degraded",
              "disabled"
            ]
          },
          "ready": {
            "type": "boolean"
          },
          "evidence": {
            "type": [
              "object",
              "null"
            ]
          },
          "evidence_expired": {
            "type": "boolean"
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "message",
              "type",
              "code"
            ],
            "properties": {
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "param": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "code": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          }
        }
      }
    }
  }
}